core: refactor client to use react-router, so it behaves like a multi page app
