+++ client/components/commititems.jsx
@@ -7,7 +7,6 @@ export default class CommitItems extends Component {
   }
 
   showCommitData(){
-    // makes data available to CommitItem Component
     return _.reduce(this.props.commitData, (accum, item)=>{
       var tagsThisItem = '';
       for(var i = 0; i < item.labels.length; i++) {
